requirements.txt
Django==5.2.7
djangorestframework==3.15.2
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1
mysqlclient==2.2.4
